refactor: move app and main file to root folder
